refactor: Member 엔티티 패키지 위치 변경
