+++ src/main/java/com/mobble/mobbleserver/refactor/meetingMember/entity/MeetingMember.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.refactor.meetingMember.entity;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.meeting.MeetingMemberErrorCode;
 import com.mobble.mobbleserver.refactor.meeting.entity.Meeting;
-import com.mobble.mobbleserver.refactor.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
