+++ src/main/java/com/mobble/mobbleserver/refactor/notification/core/entity/Notification.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.notification.core.entity;
 
 import com.mobble.mobbleserver.common.baseEntity.CreatedAtEntity;
-import com.mobble.mobbleserver.refactor.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
