+++ src/main/java/com/mobble/mobbleserver/refactor/notification/device/entity/DeviceToken.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.notification.device.entity;
 
 import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
-import com.mobble.mobbleserver.refactor.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
