+++ src/test/java/com/mobble/mobbleserver/refactor/article/repository/ArticleRepositoryImplTest.java
@@ -7,7 +7,7 @@ import com.mobble.mobbleserver.refactor.article.repository.dto.ArticleLikeInfoDt
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.like.articleLike.entity.ArticleLike;
-import com.mobble.mobbleserver.refactor.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.article.ArticleTestFixture;
 import com.mobble.mobbleserver.support.fixture.club.ClubTestFixture;
 import com.mobble.mobbleserver.support.fixture.member.MemberTestFixture;
