+++ src/test/java/com/mobble/mobbleserver/refactor/meetingMember/repository/MeetingMemberRepositoryTest.java
@@ -8,7 +8,7 @@ import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
 import com.mobble.mobbleserver.refactor.clubMember.entity.JoinStatus;
 import com.mobble.mobbleserver.refactor.meeting.entity.Meeting;
 import com.mobble.mobbleserver.refactor.meetingMember.entity.MeetingMember;
-import com.mobble.mobbleserver.refactor.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.club.ClubTestFixture;
 import com.mobble.mobbleserver.support.fixture.clubCategory.ClubCategoryTestFixture;
 import com.mobble.mobbleserver.support.fixture.clubMember.ClubMemberTestFixture;
